// src/lib/api/generated/model/assessmentsListParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export type AssessmentsListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
page_size?: number;
status?: string;
category?: string;
instrument?: string;
q?: string;
date_from?: string;
date_to?: string;
order?: string;
};